Use wrapper script to build smaller image
dockerTools.buildImage pulls in the "minimal" closure for hoogle as
determined by nix. Problem is:

1. Minimal closure is not small, because haskellPackages dependencies
   aren't very specific (e.g. docs aren't separated from other outputs).

2. It's not possible to manually prune the image in the `runAsRoot`
   phase because it read-only bind-mounts /nix/store.

So we go to a more manual method of plucking the chosen files out the
nix store and using docker to make an image from them.

Hoogle is also statically linked against Haskell libraries (but not
system libraries).